Create README.md
Include a description of what the page(s) output in a readme.txt. This is as simple as "copy and paste" the assignment requirements.
Include the HTML and CSS page name(s) in your readme.txt file.
Use good formatted, easy-to-read, and commented code.
Create a user-friendly interface, including readable output, and an interface that is self-explanatory.
The solution is written for the problem assigned and includes all required functionality.